Host API spec
Spec for Host vendor extension API, intended for the Kilo
release, though it may be backported to Juno and Icehouse.

This API exposes read-only information about the hosts in which
the midonet agents are running, to provide better visibility into
them for debugging, and to see the health of the agents.

Change-Id: I99e9256edefa0edd73f1c9d504e377e38b0d57b5